Move UITest from gallery to Issues
